Movemos textToHtml a Presenter
