+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/presentation/MoreDetailsPresenter.kt
@@ -1,9 +1,15 @@
 package ayds.lisboa.songinfo.moredetails.presentation
 
+import ayds.lisboa.songinfo.moredetails.domain.entities.Artist.ArtistData
 import ayds.lisboa.songinfo.moredetails.domain.entities.Artist
 import ayds.lisboa.songinfo.moredetails.domain.repositoryInterface.ArtistRepository
 import ayds.observer.Observable
 import ayds.observer.Subject
+import java.util.*
+
+private const val HTML_START = "<html><div width=400>"
+private const val HTML_END = "</font></div></html>"
+private const val FONT_FACE = "<font face=\"arial\">"
 
 interface MoreDetailsPresenter {
     val artistObservable: Observable<Artist>
@@ -17,9 +23,26 @@ internal class MoreDetailsPresenterImpl(private val repository: ArtistRepository
 
     override fun getArtistMoreInformation(artistName: String) {
         val artistData: Artist = repository.getArtistData(artistName)
+        artistData.geFormattingInfoArtist(artistName)
         artistObservable.notify(artistData)
     }
 
+    private fun Artist.geFormattingInfoArtist(artistName: String) {
+        val infoToFormat = (this as ArtistData).infoArtist
+        this.infoArtist = textToHtml(infoToFormat, artistName)
+    }
+
+    private fun textToHtml(text: String, term: String): String {
+        val builder = StringBuilder()
+        builder.append(HTML_START)
+        builder.append(FONT_FACE)
+        val textWithBold = text.replace("'", " ").replace("\n", "<br>").replace(
+            "(?i)$term".toRegex(), "<b>" + term.uppercase(Locale.getDefault()) + "</b>"
+        )
+        builder.append(textWithBold)
+        builder.append(HTML_END)
+        return builder.toString()
+    }
 }
